+++ src/ru/yandex/practicum/delivery/parcel/StandardParcel.java
@@ -4,7 +4,7 @@ public class StandardParcel extends Parcel {
 
     private static final int deliveryCostModifier = 2;
 
-    public StandardParcel(String description, int weight, String deliveryAddress, byte sendDay) {
+    public StandardParcel(String description, int weight, String deliveryAddress, int sendDay) {
         super(description, weight, deliveryAddress, sendDay);
     }
